% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_export_ace_output.R
\name{count_import_export_ace_uncertainty}
\alias{count_import_export_ace_uncertainty}
\title{Title}
\usage{
count_import_export_ace_uncertainty(
  tree,
  ace_nodes,
  ace_tips,
  give_tips = NA,
  start_time = NA
)
}
\arguments{
\item{tree}{#Phylogenetic tree}

\item{ace_nodes}{#Node probabilities in ace format}

\item{ace_tips}{#Tip probabilities in ace format}

\item{give_tips}{#Taxa to do the computations for}

\item{start_time}{Date of root in the phylogeny.}
}
\value{

}
\description{
Title
}
