% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitTransitionMatrix.R
\name{fitTransition_10X}
\alias{fitTransition_10X}
\title{Title}
\usage{
fitTransition_10X(tree, x, model = NA)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{x}{Tip character states that are used to inform the transition matrix}

\item{model}{Model matrix ordered where unique numbers indicate individual parameters.}
}
\value{

}
\description{
Title
}
