% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitTransitionMatrix.R
\name{FitTransitionFreeX}
\alias{FitTransitionFreeX}
\title{Title}
\usage{
FitTransitionFreeX(tree, x, model = NA, scaling_factor = 10)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{x}{Tip character states that are used to inform the transition matrix}

\item{model}{Model matrix ordered where unique numbers indicate individual parameters.}
}
\value{

}
\description{
Title
}
