% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(x)
}
\arguments{
\item{x}{The first parameter}
}
\value{
Some random text
}
\description{
We don't need any description here.
}
\examples{
hello("John")
\dontrun{
hello("Steve")
}
}
