% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importation_local_transmission.R
\name{import_local_transmission}
\alias{import_local_transmission}
\title{import_local_transmission}
\usage{
import_local_transmission(
  tree,
  LineageHomology_replicates,
  start_time,
  time_interval_size = 1/52
)
}
\arguments{
\item{time_interval_size}{}
}
\value{

}
\description{

}
