% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageHomology.R
\name{LineageHomology_w_uncertainty}
\alias{LineageHomology_w_uncertainty}
\title{LineageHomology_w_uncertainty}
\usage{
LineageHomology_w_uncertainty(
  tree,
  ace_nodes,
  ace_tips,
  give_tips = NA,
  start_time = NA
)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{ace_nodes}{Node probabilities in ace format}

\item{ace_tips}{Tip probabilities in ace format}

\item{give_tips}{Constrain the computations to a set of tips, e.g. all the names of the tips in a certain state (e.g. one specific geographical location).}

\item{start_time}{Date of root in the phylogeny.}
}
\value{
See LineageHomology for descriptions of the returns.
}
\description{
LineageHomology_w_uncertainty returns the same outputs as LineageHomology.
 LineageHomology counts transmission lineages according to state transitions between ancestral and descendant nodes that have a probability higher than 50 percent.
 LineageHomolog_w_uncertainty instead samples states from the posterior probability of states for each node.
 Thus, the transmission lineage division is probabilistic and retains the uncertainty in the posterior distribution if one runs multiple times,
  e.g., using base::replicate() (see examples).
}
\examples{
\dontrun{
multi_counts = pbreplicate(
  100,
  LineageHomology::LineageHomology_w_uncertainty(
    tree,
    ace_nodes = test$lik.anc,
    give_tips = Norwegian_tips,
    ace_tips = to.matrix(Locations, seq = c("Norway", "RoW")),
    start_time = start_date
  )
)
}
}
