% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageHomology.R
\name{LineageHomology_w_uncertainty_test}
\alias{LineageHomology_w_uncertainty_test}
\title{Title}
\usage{
LineageHomology_w_uncertainty_test(
  tree,
  ace_nodes,
  ace_tips,
  give_tips = NA,
  start_time = NA
)
}
\arguments{
\item{start_time}{}
}
\value{

}
\description{
Title
}
