% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageHomology.R
\name{LineageHomology_w_uncertainty_v2}
\alias{LineageHomology_w_uncertainty_v2}
\title{LineageHomology_w_uncertainty_v2}
\usage{
LineageHomology_w_uncertainty_v2(
  tree,
  ace_nodes,
  ace_tips,
  give_tips = NA,
  start_time = NA
)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{ace_nodes}{Node probabilities in ace format}

\item{ace_tips}{Tip probabilities in ace format}

\item{give_tips}{Taxa to do the computations for}

\item{start_time}{Date of root in the phylogeny.}
}
\value{

}
\description{
LineageHomology_w_uncertainty_v2 returns the same outputs as LineageHomology.
 This is the same function as LineageHomology_w_uncertainty but faster.
 LineageHomology counts transmission lineages according to state transitions between ancestral and descendant nodes that have a probability higher than 50 percent.
 LineageHomolog_w_uncertainty instead samples states from the posterior probability of states (that is usually included in a phylogeographical consensus tree) for each node. Thus, the transmission lineage division is probabilistic and retains the uncertainty in the posterior distribution if the function is run multiple times, e.g., using base::replicate().
}
\examples{
\dontrun{
multi_counts = pbreplicate(
  100,
  LineageHomology::LineageHomology_w_uncertainty_v2(
    tree,
    ace_nodes = test$lik.anc,
    give_tips = Norwegian_tips,
    ace_tips = to.matrix(Locations, seq = c("Norway", "RoW")),
    start_time = start_date
  )
)
}

}
