% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitTransitionMatrix.R
\name{makeQ10X}
\alias{makeQ10X}
\title{Title}
\usage{
makeQ10X(m, q, index.matrix)
}
\arguments{
\item{m}{number of character states}

\item{q}{number of param}

\item{index.matrix}{}
}
\value{
Transition matrix in the correct form for Felsensteins pruning algorithm.
}
\description{
Title
}
