% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitTransitionMatrix.R
\name{makeQfreeX}
\alias{makeQfreeX}
\title{Title}
\usage{
makeQfreeX(m, q, index.matrix, scale)
}
\arguments{
\item{m}{number of character states}

\item{q}{number of param}

\item{scale}{scaling of the transition matrix in and out parameters.}
}
\value{
Transition matrix in the correct form for Felsensteins pruning algorithm.
}
\description{
Title
}
