% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nodepath_quick_w_mrca}
\alias{nodepath_quick_w_mrca}
\title{Written by Magnus Nygaard Osnes 2021.}
\usage{
nodepath_quick_w_mrca(tree, taxa, mrca)
}
\arguments{
\item{mrca}{}
}
\value{

}
\description{
Written by Magnus Nygaard Osnes 2021.
}
