% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{nodepath_quick_w_mrca_nodenumber}
\alias{nodepath_quick_w_mrca_nodenumber}
\title{nodepath quick
Written by Magnus Nygaard Osnes 2021.}
\usage{
nodepath_quick_w_mrca_nodenumber(tree, taxa, mrca)
}
\arguments{
\item{mrca}{}
}
\value{

}
\description{
nodepath quick
Written by Magnus Nygaard Osnes 2021.
}
