% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_export_local_transmission.R
\name{plot_import_export_local_transmission}
\alias{plot_import_export_local_transmission}
\title{plot_import_export_local_transmission}
\usage{
plot_import_export_local_transmission(
  tree,
  result_import_export_local_transmission,
  time_interval_size = 1/52,
  start_time,
  date_breaks = 2,
  time_interval = c("2000-05-01", "2016-05-01"),
  main_title = "",
  text_size = 10
)
}
\arguments{
\item{tree}{A tree object for which the geographical estimation has been done.}

\item{result_import_export_local_transmission}{A list object containing the results from running the import_local_transmission function.}

\item{time_interval_size}{A numeric value representing the size of the time intervals used in import_local_transmission.}

\item{start_time}{A date object representing the root of the phylogeny.}

\item{date_breaks}{A string representing the interval between dates plotted on the x-axis. Acceptable arguments include "1 week", "2 months","1 year" etc.}

\item{importation_or_local}{A string argument that specifies whether to plot importation, local transmission or both. Acceptable arguments are: "importation","local" or "both".
Note: For plotting both importation and local transmission, the grid r-package is required.}
}
\value{
This function returns a plot of importation and/or local transmission events over time.
}
\description{
This function plots the estimates of import, export and local transmission events that are obtained from running the import_export_local_transmission function. It takes the tree for which the geographical estimation has been done, the results from running import_local_transmission, the size of the time intervals used in import_local_transmission, the date of the root of the phylogeny, and the interval between dates plotted on the x-axis as input arguments.
}
