% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importation_local_transmission.R
\name{plot_importation_local_transmission}
\alias{plot_importation_local_transmission}
\title{Title}
\usage{
plot_importation_local_transmission(
  tree,
  result_import_local_transmission,
  time_interval_size = 1/52,
  start_time,
  date_breaks = "1 month",
  importation_or_local = "both"
)
}
\arguments{
\item{importation_or_local}{}
}
\value{

}
\description{
Title
}
