% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relative_load_multi.R
\name{Relative_load_import_multi}
\alias{Relative_load_import_multi}
\title{Title}
\usage{
Relative_load_import_multi(tree, multicount, dates, start_time, tip_list)
}
\arguments{
\item{tree}{Takes the phylogenetic tree used for ancestral state reconstruction.}

\item{multicount}{Takes the Result_multi output from and replicated count script on castor (which runs on ace, simmap etc.)}

\item{dates}{Takes argument dates, should have names matching tip labels from tree.}

\item{start_time}{Takes the date of the root of the phylogenetic tree.}

\item{tip_list}{Takes a list of taxa which to compute the importation load for: e.g. all tips in Norway}
}
\value{

}
\description{
Title
}
