% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{reorder_LH}
\alias{reorder_LH}
\title{reorder_LH}
\usage{
reorder_LH(result_LH)
}
\arguments{
\item{result_LH}{#Return from LineageHomology.}
}
\value{

}
\description{
Takes the output of LineageHomology and reorders the transmission lineages so that the largest comes first 1, second largest comes 2 etc.
}
