% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_export_local_transmission.R
\name{Summarize_import_export_local_transmission}
\alias{Summarize_import_export_local_transmission}
\title{Summarize_import_export_local_transmission}
\usage{
Summarize_import_export_local_transmission(
  result_import_export_local_transmission
)
}
\arguments{
\item{LineageHomology_replicates}{Results from running LineageHomology_w_uncertainty_v2}
}
\value{
The function returns quantiles of the 2.5, 50, and 97.5 percent quantiles of the observed number of importations, local transmissions and exports.
}
\description{
This function summarizes the estimates of import, export and local transmission events from
running import_export_local_transmission. Keep in mind that both import and local transmission events can produce export events.
}
