% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importation_local_transmission.R
\name{Summarize_import_local_transmission}
\alias{Summarize_import_local_transmission}
\title{Title}
\usage{
Summarize_import_local_transmission(LineageHomology_replicates)
}
\arguments{
\item{LineageHomology_replicates}{}
}
\value{

}
\description{
Title
}
