% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_plots.R
\name{treemap_lineagehomology}
\alias{treemap_lineagehomology}
\title{Title}
\usage{
treemap_lineagehomology(Result_LineageHomology)
}
\arguments{
\item{Result_LineageHomology}{Result from the function LineageHomology}
}
\value{

}
\description{
Title
}
