% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_plots.R
\name{lineage_info}
\alias{lineage_info}
\title{Title}
\usage{
lineage_info(Result_Lineage_homology, name_date)
}
\arguments{
\item{Result_Lineage_homology}{The output from linege homology.}

\item{name_date}{#Data frame with taxa names in one column and numeric dates of sampling in another. The column names should be c("name", "date")}
}
\value{

}
\description{
Title
}
