% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineageHomology.R
\name{LineageHomology_w_uncertainty}
\alias{LineageHomology_w_uncertainty}
\title{Title}
\usage{
LineageHomology_w_uncertainty(
  tree,
  ace_nodes,
  ace_tips,
  give_tips = NA,
  start_time = NA
)
}
\arguments{
\item{tree}{#Phylogenetic tree}

\item{ace_nodes}{#Node probabilities in ace format}

\item{ace_tips}{#Tip probabilities in ace format}

\item{give_tips}{#Taxa to do the computations for}

\item{start_time}{Date of root in the phylogeny.}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
#LineageHomology_w_uncertainty(tree_test, ace_nodes=fit1$lik.anc,
ace_tips = to.matrix(loc, seq=c("Norway", "RoW")), start_time=2000)
}
}
