% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_plots.R
\name{ridgeplot_lineagedensities}
\alias{ridgeplot_lineagedensities}
\title{Title}
\usage{
ridgeplot_lineagedensities(
  Result_lineage_info,
  datelims,
  groups_larger_than = 4,
  color_by_state = FALSE
)
}
\arguments{
\item{datelims}{}
}
\value{

}
\description{
Title
}
