% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_plots.R
\name{treemap_lineagehomology}
\alias{treemap_lineagehomology}
\title{treemap_lineagehomology}
\usage{
treemap_lineagehomology(Result_LineageHomology)
}
\arguments{
\item{Result_LineageHomology}{Result from the function LineageHomology}
}
\value{

}
\description{
Treemap_lineagehomology uses the treemap package to visualize transmission lineages (TL) as rectangles with areas that reflect their relative sizes.
 Additionally, the text on each square gives the group number of the TL,
 the estimated time of the most recent common ancestor (TMRCA) and the number of tips (S) in the TL. See the full tutorial at
 \url{https://github.com/magnusnosnes/LineageHomology/blob/master/Examples_and_plotting_methods/Simple_example/Basic_plotting.md}
}
